# docs/src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is HexLabs' API Documentation.",
    "title": "API Reference",
    "version": "1.0.0"
  },
...
  "schemes": ["https", "http"],
  "paths": {
...
    "/applications": {
      "get": {
        "tags": ["registration"],
        "summary": "Finds and displays information of all applications.",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "hexathon",
            "in": "query",
            "description": "Filter applications by hexathon id",
            "type": "string"
          },
          {
            "name": "userId",
            "in": "query",
            "description": "Filter applications by user id",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Application"
              }
            }
          },
          "400": {
            "description": "Invalid status"
          }
        }
      }
    },
    "/applications/{id}": {
      "get": {
        "tags": ["registration"],
        "summary": "Finds and displays application by id.",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Application id that needs to be considered for filter.",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Application"
            }
          },
          "400": {
            "description": "Invalid status"
          }
        }
      }
    },
    "/applications/actions/choose-application-branch": {
      "post": {
        "tags": ["registration"],
        "summary": "Creates a new application based on the branch chosen.",
        "description": "",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Request body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "hexathon": {
                  "type": "string"
                },
                "applicationBranch": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Application"
            }
          },
          "400": {
            "description": "Invalid Status"
          }
        }
      }
    },
    "/applications/{id}/actions/save-application-data": {
      "post": {
        "tags": ["registration"],
        "summary": "Saves application data.",
        "description": "",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Application id.",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "description": "Request body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "applicationData": {
                  "type": "object"
                },
                "branchFormPage": {
                  "type": "number"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Application"
            }
          },
          "400": {
            "description": "Invalid Status"
          }
        }
      }
    },
    "/applications/{id}/actions/submit-application": {
      "post": {
        "tags": ["registration"],
        "summary": "Submit an application.",
        "description": "",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Application id.",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Application"
            }
          },
          "400": {
            "description": "Invalid Status"
          }
        }
      }
    },
    "/applications/{id}/actions/save-confirmation-data": {
      "post": {
        "tags": ["registration"],
        "summary": "Saves confirmation data.",
        "description": "",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Application id.",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "description": "Request body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "confirmationData": {
                  "type": "object"
                },
                "branchFormPage": {
                  "type": "number"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Application"
            }
          },
          "400": {
            "description": "Invalid Status"
          }
        }
      }
    },
    "/applications/{id}/actions/submit-confirmation": {
      "post": {
        "tags": ["registration"],
        "summary": "Submit a confirmation.",
        "description": "",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "Application id.",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Application"
            }
          },
          "400": {
            "description": "Invalid Status"
          }
        }
      }
    },
...
    "/users/{userId}": {
      "get": {
        "tags": ["users"],
        "summary": "Finds and displays profile by user ID.",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "type": "integer",
            "description": "User ID",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "allOf": [{ "$ref": "#/definitions/Profile" }, { "$ref": "#/definitions/Permission" }]
            }
          },
          "400": {
            "description": "Invalid Status"
          }
        }
      },
      "put": {
        "tags": ["users"],
        "summary": "Update a user's profile.",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "type": "integer",
            "description": "User ID",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Profile"
            }
          },
          "400": {
            "description": "Invalid Status"
          }
        }
      }
    },
    "/users/actions/retrieve": {
      "post": {
        "tags": ["users"],
        "summary": "Gets user profiles by list of userIds",
        "description": "",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "Request body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "userIds": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Profile"
              }
            }
          },
          "400": {
            "description": "Invalid Status"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Application": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "hexathon": {
          "type": "integer"
        },
        "applicationBranch": {
          "$ref": "#/definitions/Branch"
        },
        "applicationStartTime": {
          "type": "string",
          "format": "date-time"
        },
        "applicationSubmitTime": {
          "type": "string",
          "format": "date-time"
        },
        "applicationData": {
          "$ref": "#/definitions/ApplicationData"
        },
        "confirmationBranch": {
          "$ref": "#/definitions/Branch"
        },
        "confirmationStartTime": {
          "type": "string",
          "format": "date-time"
        },
        "confirmationSubmitTime": {
          "type": "string",
          "format": "date-time"
        },
        "confirmationData": {
          "type": "object"
        },
        "status": {
          "type": "string",
          "enum": ["DRAFT", "APPLIED", "ACCEPTED", "CONFIRMED", "DENIED"]
        },
        "gradingComplete": {
          "type": "number"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "xml": {
        "name": "Application"
      }
    },
    "ApplicationData": {
      "type": "object",
      "properties": {
        "adult": {
          "type": "boolean"
        },
        "occupation": {
          "type": "string"
        },
        "school": {
          "type": "string"
        },
        "schoolEmail": {
          "type": "string"
        },
        "schoolYear": {
          "type": "string"
        },
        "graduationYear": {
          "type": "number"
        },
        "major": {
          "type": "string"
        },
        "shirtSize": {
          "type": "string"
        },
        "dietaryRestriction": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "ethnicity": {
          "type": "string"
        },
        "address": {
          "type": "object",
          "properties": {
            "line1": {
              "type": "string"
            },
            "line2": {
              "type": "string"
            },
            "city": {
              "type": "string"
            },
            "state": {
              "type": "string"
            },
            "zip": {
              "type": "number"
            },
            "country": {
              "type": "string"
            }
          }
        },
        "marketing": {
          "type": "string"
        },
        "website": {
          "type": "string"
        },
        "linknedin": {
          "type": "string"
        },
        "extraInfo": {
          "type": "string"
        },
        "confirmChecks": {
          "type": "object"
        },
        "essays": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "criteria": {
                "type": "string"
              },
              "answer": {
                "type": "string"
              }
            }
          }
        }
      },
      "xml": {
        "name": "ApplicationData"
      }
    },
    "Branch": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "hexathon": {
          "type": "integer"
        },
        "type": {
          "type": "string",
          "enum": ["APPLICATION", "CONFIRMATION"]
        },
        "settings": {
          "type": "object",
          "properties": {
            "open": {
              "type": "string",
              "format": "date-time"
            },
            "closed": {
              "type": "string",
              "format": "date-time"
            }
          }
        },
        "formPages": {
          "type": "array",
          "items": {
            "properties": {
              "title": {
                "type": "string"
              },
              "jsonSchema": {
                "type": "object",
                "default": {}
              },
              "uiSchema": {
                "type": "object",
                "default": {}
              }
            }
          }
        }
      },
      "xml": {
        "name": "Branch"
      }
    },
...
    "Profile": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "name": {
          "type": "object",
          "properties": {
            "first": {
              "type": "string"
            },
            "middle": {
              "type": "string"
            },
            "last": {
              "type": "string"
            }
          }
        },
        "phoneNumber": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "resume": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Profile"
      }
    },
    "Permission": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "roles": {
          "type": "object",
          "properties": {
            "member": {
              "type": "boolean"
            },
            "exec": {
              "type": "boolean"
            },
            "admin": {
              "type": "boolean"
            }
          }
        }
      },
      "xml": {
        "name": "Permission"
      }
    },
...
  }
}